+++ src/browser-utilities.ts
@@ -137,6 +137,15 @@ export function initBrowserUtilities(options: BrowserUtilitiesOptions) {
         }
     }
 
+    let spawnReindeer = (spawnLead?: boolean) => {
+        // TODO: remove the "any" as soon as the typo in @types/cookieclicker is fixed
+        let reindeer = new (Game as any).shimmer('reindeer') as Game.Shimmer;
+        if(typeof spawnLead != 'boolean' || spawnLead) {
+            reindeer.spawnLead = 1;
+        }
+        return reindeer;
+    }
+
     window.CConnoisseur = {
         mockedDate,
         clearNewsTickerText,
@@ -147,5 +156,6 @@ export function initBrowserUtilities(options: BrowserUtilitiesOptions) {
         reincarnate,
         redrawMarketMinigame,
         startGrandmapocalypse,
+        spawnReindeer,
     };
 }