+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added: Browser utilities `CConnoisseur.ascend` and `CConnoisseur.reincarnate`.
 - Added: Browser utility `CConnoisseur.redrawBankMinigame`.
 - Added: Browser utility `CConnoisseur.startGrandmapocalypse`.
+- Added: Browser utility `CConnoisseur.spawnReindeer`.
 
 ## 0.2.0 - 2021-10-01
 - Added: Explicit support for Firefox.